chore: move faqData into dedicated file
