+++ src/pages/Contact.tsx
@@ -31,6 +31,7 @@ import MarquesZahir from '../assets/Marques-Zahir.png'; // Replace with actual i
 import CaseyBass from '../assets/Casey-Bass.png'; // Replace with actual image path
 import AshleyLewis from '../assets/Ashley-Lewis.png'; // Replace with actual image path
 import connectTeam from '../assets/connect-team.png'; // Replace with actual image path
+import faqData from '../constants/faq';
 
 const Contact: React.FC = () => {
   const [activeIndexes, setActiveIndexes] = useState<{
@@ -108,51 +109,6 @@ const Contact: React.FC = () => {
       );
     };
   }, []);
-
-  const faqData = [
-    {
-      questions: [
-        'What is your return policy?',
-        'How do I track my order?',
-        'Can I purchase items in bulk?',
-        'Do you ship internationally?',
-      ],
-      answers: [
-        'Our return policy allows returns within 30 days of purchase.',
-        'You can track your order using the tracking number provided in your email.',
-        'Yes, we offer bulk purchasing options. Please contact our support team for more details.',
-        'Yes, we ship to many countries around the world. Shipping charges may vary.',
-      ],
-    },
-    {
-      questions: [
-        'How can I reset my password?',
-        'Where is your company located?',
-        'Do you have customer support?',
-        'What payment methods do you accept?',
-      ],
-      answers: [
-        'You can reset your password by clicking on the "Forgot Password" link on the login page.',
-        'Our company is located at 123 Main Street, Anytown, USA.',
-        'Yes, we offer 24/7 customer support. You can contact us via email or phone.',
-        'We accept various payment methods including credit cards, PayPal, and bank transfers.',
-      ],
-    },
-    {
-      questions: [
-        'What are your business hours?',
-        'How do I make a complaint?',
-        'Do you offer gift cards?',
-        'How can I apply for a job?',
-      ],
-      answers: [
-        'Our business hours are from 9 AM to 6 PM, Monday to Friday.',
-        'You can make a complaint by contacting our customer support team via email.',
-        'Yes, we offer gift cards. You can purchase them on our website.',
-        'You can apply for a job by visiting our careers page and submitting your resume.',
-      ],
-    },
-  ];
 
   return (
     <>
